calculator/math-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sympy==1.12
python-dotenv==1.0.0
google-genai>=1.35.0